+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -28,6 +28,8 @@ public protocol HIGTheme: Sendable {
     var steps: any HIGStepsTokens { get }
     var pearlSteps: any HIGPearlStepsTokens { get }
     var timeline: any HIGTimelineTokens { get }
+    var statusIndicator: any HIGStatusIndicatorTokens { get }
+    var emptyState: any HIGEmptyStateTokens { get }
     var progress: any HIGProgressTokens { get }
     var alert: any HIGAlertTokens { get }
     var toast: any HIGToastTokens { get }